# src/jvmMain/resources/api/SyncService_OpenApi_1.0.0.yaml
openapi: 3.0.3
info:
  title: Scorewarrior synchronise service
  description: API for tool that streamlines interaction of game designers and artists
  version: 1.0.0
servers:
  - url: 'https://scorewarrior.sync.service.com/v1'
  - url: 'http://scorewarrior.sync.service.com/v1'
tags:
  - name: Element
    description: Information about heroes and weapons
paths:
  '/element/{type}':
    post:
      tags:
        - element
      summary: Add a new element profile
      operationId: addElement
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
      requestBody:
        description: Add a new element profile
        content:
          multipart/form-data:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Hero'
                - $ref: '#/components/schemas/Weapon'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    type: string
                    description: Link on the created element
        '405':
          description: Invalid input
          content: { }
      x-codegen-request-body-name: body
    get:
      tags:
        - element
      summary: Find all existing entities
      operationId: getAllElements
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/Hero'
                    - $ref: '#/components/schemas/Weapon'
        '204':
          description: No any element found
          content: { }
    put:
      tags:
        - element
      summary: Update an existing element profile
      operationId: updateElement
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
      requestBody:
        description: Update profile
        content:
          multipart/form-data:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Hero'
                - $ref: '#/components/schemas/Weapon'
        required: true
      responses:
        '200':
          description: Updated
          content: { }
        '404':
          description: Element not found
          content: { }
        '405':
          description: Validation exception
          content: { }
      x-codegen-request-body-name: body
  '/element/{type}/{name}':
    get:
      tags:
        - element
      summary: Get element by name
      description: Returns a single element
      operationId: getElementByName
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Element name to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Hero'
                  - $ref: '#/components/schemas/Weapon'

        '400':
          description: Invalid name supplied
          content: { }
        '404':
          description: Element not found
          content: { }

  '/element/{type}/{name}/{user_id}':
    get:
      tags:
        - element
      summary: Get draft element by name
      description: Returns a single element
      operationId: getDraftElementByName
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Element name to return
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          description: Element name to return
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Hero'
                  - $ref: '#/components/schemas/Weapon'

        '400':
          description: Invalid name supplied
          content: { }
        '404':
          description: Element not found
          content: { }
    delete:
      tags:
        - element
      summary: Deletes an element profile
      operationId: deleteDraftElement
      parameters:
        - name: type
          in: path
          description: Type of element to delete
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Element name to delete
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          description: Element name to return
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Deleted
          content: { }
        '404':
          description: Element not found
          content: { }

  '/element/{type}/{user_id}':
    put:
      tags:
        - element
      summary: Update an existing draft element profile
      operationId: updateDraftElement
      parameters:
        - name: type
          in: path
          description: Type of element to return
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          description: Element name to return
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        description: Update profile
        content:
          multipart/form-data:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Hero'
                - $ref: '#/components/schemas/Weapon'
        required: true
      responses:
        '200':
          description: Updated
          content: { }
        '404':
          description: Element not found
          content: { }
        '405':
          description: Validation exception
          content: { }
      x-codegen-request-body-name: body
components:
  schemas:
    Weapon:
      required:
        - name
      type: object
      properties:
        name:
          description: Name of weapon
          type: string
          example: Sword
          minLength: 1
          maxLength: 100
        mainImage:
          description: Link to main image in filesystem
          type: string
          format: binary
          minLength: 1
          maxLength: 200
        entireIcon:
          description: Link to entire icon in filesystem
          type: string
          format: binary
          minLength: 1
          maxLength: 200
        brokenIcon:
          description: Link to broken icon in filesystem
          type: string
          format: binary
          minLength: 1
          maxLength: 200

      xml:
        name: Weapon
    Hero:
      required:
        - name
      type: object
      properties:
        name:
          description: Name of hero
          type: string
          example: Packman
          minLength: 1
          maxLength: 100
        mainImage:
          description: Link to main image in filesystem
          type: string
          format: binary
          minLength: 1
          maxLength: 200
        icon:
          description: Link to icon in filesystem
          type: string
          format: binary
          minLength: 1
          maxLength: 200
      xml:
        name: Hero